Folder structure changes and cross-check scripts
