+++ Makefile
@@ -5,7 +5,7 @@
 
 # Usage: make all
 # Creates the EDA report and final analysis report through checking all dependencies
-all: src/crime_vancouver_eda.md doc/vancouver_crime_predict_report.md
+all: src/report-eda/crime_vancouver_eda.md doc/vancouver_crime_predict_report.md
 
 # Usage: make analysis
 # Runs the analysis except rendering the report
@@ -13,12 +13,12 @@ analysis: data/processed/test_target.csv src/figure-eda/crime_top5.png results/p
 
 # Usage: make report
 # Renders the EDA report and final report
-report: src/crime_vancouver_eda.md doc/vancouver_crime_predict_report.md
+report: src/report-eda/crime_vancouver_eda.md doc/vancouver_crime_predict_report.md
 
 # Usage: make data/raw/crimedata_csv_all_years.csv
 # Downloads the necessary crime data from Vancouver Police Department and extract the zip file
-data/raw/crimedata_csv_all_years.csv: src/download_data.py
-	python src/download_data.py --url=https://geodash.vpd.ca/opendata/crimedata_download/crimedata_csv_all_years.zip?disclaimer=on --file_path=data/raw --zip_file_name=crimedata_csv_all_years.csv
+data/raw/crimedata_csv_all_years.csv: src/scripts/download_data.py
+	python src/scripts/download_data.py --url=https://geodash.vpd.ca/opendata/crimedata_download/crimedata_csv_all_years.zip?disclaimer=on --file_path=data/raw --zip_file_name=crimedata_csv_all_years.csv
 
 # Usage: make data/processed/training_feature.csv
 # Date range can be specified in the script "python src/..." with from_year and to_year parameter.
@@ -30,8 +30,8 @@ data/processed/test_target.csv \
 data/processed/training_df.csv \
 src/figure-preprocess/data_before_balance.png \
 src/figure-preprocess/data_after_balance.png \
-src/figure-preprocess/observations.png: src/split_data.py data/raw/crimedata_csv_all_years.csv
-	python src/split_data.py --input_path=data/raw/crimedata_csv_all_years.csv --out_path=data/processed/  --graph_path=src/figure-preprocess/ --from_year=2016 --to_year=2020
+src/figure-preprocess/observations.png: src/scripts/split_data.py data/raw/crimedata_csv_all_years.csv
+	python src/scripts/split_data.py --input_path=data/raw/crimedata_csv_all_years.csv --out_path=data/processed/  --graph_path=src/figure-preprocess/ --from_year=2016 --to_year=2020
 
 # Usage: make src/figure-eda/neighbour_crimes.png
 # Performs EDA and generates necessary graphs and tables
@@ -39,18 +39,18 @@ src/figure-eda/neighbour_crimes.png \
 src/figure-eda/crime_type.png \
 src/figure-eda/crime_evolution.png \
 src/figure-eda/crime_correlation.png \
-src/figure-eda/crime_top5.png: src/eda_script.py data/processed/training_df.csv
-	python src/eda_script.py --input_path=data/processed/training_df.csv --out_dir=src/figure-eda/ 
+src/figure-eda/crime_top5.png: src/scripts/eda_script.py data/processed/training_df.csv
+	python src/scripts/eda_script.py --input_path=data/processed/training_df.csv --out_dir=src/figure-eda/ 
 
 # Usage: make src/crime_vancouver_eda.md
 # Render EDA report
-src/crime_vancouver_eda.md: src/crime_vancouver_eda.Rmd src/figure-eda/neighbour_crimes.png src/figure-preprocess/observations.png src/figure-eda/crime_type.png src/figure-eda/crime_evolution.png src/figure-eda/crime_correlation.png src/figure-eda/crime_top5.png
-	Rscript -e "rmarkdown::render('src/crime_vancouver_eda.Rmd')"
+src/report-eda/crime_vancouver_eda.md: src/report-eda/crime_vancouver_eda.Rmd src/figure-eda/neighbour_crimes.png src/figure-preprocess/observations.png src/figure-eda/crime_type.png src/figure-eda/crime_evolution.png src/figure-eda/crime_correlation.png src/figure-eda/crime_top5.png
+	Rscript -e "rmarkdown::render('src/report-eda/crime_vancouver_eda.Rmd')"
 
 # Usage: make data/processed/models.p
 # create pre-processor for column transformation
-data/processed/preprocessor.p data/processed/models.p: src/pre_process_data.py
-	python src/pre_process_data.py --out_path=data/processed/
+data/processed/preprocessor.p data/processed/models.p: src/scripts/pre_process_data.py
+	python src/scripts/pre_process_data.py --out_path=data/processed/
 
 # Usage: make results/pipe_best.p
 # Fit and tune model
@@ -58,14 +58,14 @@ results/pipe_best.p \
 results/models_results_cv.png \
 results/best_LR_model.png \
 results/confusion_matrix.png \
-results/classification_report.png: src/modelling.py \
+results/classification_report.png: src/scripts/modelling.py \
 data/processed/training_feature.csv \
 data/processed/training_target.csv \
 data/processed/test_feature.csv \
 data/processed/test_target.csv \
 data/processed/models.p \
 data/processed/preprocessor.p
-	python src/modelling.py --input_path=data/processed/ --out_path=results/
+	python src/scripts/modelling.py --input_path=data/processed/ --out_path=results/
 
 # Usage: make doc/vancouver_crime_predict_report.md
 # render final report
@@ -87,5 +87,5 @@ clean:
 	rm -rf data/raw/crimedata_csv_all_years.csv
 	rm -rf results
 	rm -rf src/figure-eda src/figure-preprocess
-	rm -rf src/crime_vancouver_eda.md src/crime_vancouver_eda.html	
+	rm -rf src/report-eda/crime_vancouver_eda.md src/report-eda/crime_vancouver_eda.html	
 	rm -rf doc/vancouver_crime_predict_report.md doc/vancouver_crime_predict_report.html
