+++ doc/vancouver_crime_predict_report.md
@@ -116,9 +116,9 @@ Figure 4. The data after rebalancing
 
 Prior to model fitting, the data was preprocessed via column
 transformers. SimpleImputer and OneHotEncoder were applied to the
-categorical feature, “Neighbourhood.” OneHotEncoding was used for the
-date components, “Year,” “Month,” “Hour” and “Minute.” And finally,
-SimpleImputer and StandardScaler were applied to “X”and “Y,” which are
+categorical feature, “Neighbourhood”. OneHotEncoding was used for the
+date components, “Year”, “Month”, “Hour” and “Minute”. And finally,
+SimpleImputer and StandardScaler were applied to “X”and “Y”, which are
 coordinate points of the crime location.
 
 ### Finding the best performing model
