+++ src/figure-makefile/Makefile.dot
@@ -1,0 +1,109 @@
+digraph G {
+n2[label="all", color="red"];
+n27[label="data/processed/models.p", color="red"];
+n29[label="data/processed/preprocessor.p", color="red"];
+n25[label="data/processed/test_feature.csv", color="red"];
+n26[label="data/processed/test_target.csv", color="red"];
+n7[label="data/processed/training_df.csv", color="red"];
+n23[label="data/processed/training_feature.csv", color="red"];
+n24[label="data/processed/training_target.csv", color="red"];
+n9[label="data/raw/crimedata_csv_all_years.csv", color="red"];
+n18[label="doc/references.bib", color="green"];
+n17[label="doc/vancouver_crime_predict_report.Rmd", color="green"];
+n16[label="doc/vancouver_crime_predict_report.md", color="red"];
+n30[label="results/best_LR_model.png", color="red"];
+n32[label="results/classification_report.png", color="red"];
+n31[label="results/confusion_matrix.png", color="red"];
+n21[label="results/models_results_cv.png", color="red"];
+n4[label="src/crime_vancouver_eda.Rmd", color="green"];
+n3[label="src/crime_vancouver_eda.md", color="red"];
+n6[label="src/crime_vancouver_eda.py", color="green"];
+n10[label="src/download_data.py", color="green"];
+n14[label="src/figure-eda/crime_correlation.png", color="red"];
+n13[label="src/figure-eda/crime_evolution.png", color="red"];
+n15[label="src/figure-eda/crime_top5.png", color="red"];
+n12[label="src/figure-eda/crime_type.png", color="red"];
+n5[label="src/figure-eda/neighbour_crimes.png", color="red"];
+n20[label="src/figure-preprocess/data_after_balance.png", color="red"];
+n19[label="src/figure-preprocess/data_before_balance.png", color="red"];
+n11[label="src/figure-preprocess/observations.png", color="red"];
+n22[label="src/modelling.py", color="green"];
+n28[label="src/pre_process_data.py", color="green"];
+n8[label="src/split_data.py", color="green"];
+n16 -> n2 ; 
+n3 -> n2 ; 
+n28 -> n27 ; 
+n28 -> n29 ; 
+n9 -> n25 ; 
+n8 -> n25 ; 
+n9 -> n26 ; 
+n8 -> n26 ; 
+n9 -> n7 ; 
+n8 -> n7 ; 
+n9 -> n23 ; 
+n8 -> n23 ; 
+n9 -> n24 ; 
+n8 -> n24 ; 
+n10 -> n9 ; 
+n18 -> n16 ; 
+n17 -> n16 ; 
+n30 -> n16 ; 
+n32 -> n16 ; 
+n31 -> n16 ; 
+n21 -> n16 ; 
+n14 -> n16 ; 
+n15 -> n16 ; 
+n20 -> n16 ; 
+n19 -> n16 ; 
+n27 -> n30 ; 
+n29 -> n30 ; 
+n25 -> n30 ; 
+n26 -> n30 ; 
+n23 -> n30 ; 
+n24 -> n30 ; 
+n22 -> n30 ; 
+n27 -> n32 ; 
+n29 -> n32 ; 
+n25 -> n32 ; 
+n26 -> n32 ; 
+n23 -> n32 ; 
+n24 -> n32 ; 
+n22 -> n32 ; 
+n27 -> n31 ; 
+n29 -> n31 ; 
+n25 -> n31 ; 
+n26 -> n31 ; 
+n23 -> n31 ; 
+n24 -> n31 ; 
+n22 -> n31 ; 
+n27 -> n21 ; 
+n29 -> n21 ; 
+n25 -> n21 ; 
+n26 -> n21 ; 
+n23 -> n21 ; 
+n24 -> n21 ; 
+n22 -> n21 ; 
+n4 -> n3 ; 
+n14 -> n3 ; 
+n13 -> n3 ; 
+n15 -> n3 ; 
+n12 -> n3 ; 
+n5 -> n3 ; 
+n11 -> n3 ; 
+n7 -> n14 ; 
+n6 -> n14 ; 
+n7 -> n13 ; 
+n6 -> n13 ; 
+n7 -> n15 ; 
+n6 -> n15 ; 
+n7 -> n12 ; 
+n6 -> n12 ; 
+n7 -> n5 ; 
+n6 -> n5 ; 
+n9 -> n20 ; 
+n8 -> n20 ; 
+n9 -> n19 ; 
+n8 -> n19 ; 
+n9 -> n11 ; 
+n8 -> n11 ; 
+}
